// src/app/core/models/detalleFactura.model.ts
export interface DetalleFactura {
  idPRODUCTO: string;
  txtITEM: string;
  txtUNIDAD_MEDIDA_DET: string;
  txtCANTIDAD_DET: string;
  txtPRECIO_DET: string;
  txtSUB_TOTAL_DET: string;
  txtPRECIO_TIPO_CODIGO: string;
  txtIGV: string;
  txtISC: string;
  txtIMPORTE_DET: string;
  txtCOD_TIPO_OPERACION: string;
  txtCODIGO_DET: string;
  txtDESCRIPCION_DET: string;
  txtPRECIO_SIN_IGV_DET: string;
}